create a library for helper functions
